Use abstract_atomic.h
Remove the atomic functions from cache_inode.h and update consumers to
use the new functions in abstract_atomic.h

+++ src/include/cache_inode_lru.h
@@ -159,7 +159,7 @@ static const uint32_t LRU_NO_LANE = ~0;
 extern void cache_inode_lru_pkginit(void);
 extern void cache_inode_lru_pkgshutdown(void);
 
-extern uint32_t open_fd_count;
+extern size_t open_fd_count;
 
 extern struct cache_entry_t *cache_inode_lru_get(cache_inode_status_t *status,
                                                  uint32_t flags);
